Updated the way posts were being initialized
Previously posts were being initialized in stages. The metadata was
parsed when the instance was first created and the markdown would be
read in and converted to HTML only when needed. This was causing some
headache, so now everything is done on instantiation. Also, the
metadata is now separated from the content by a single line containing
three dashes (---) rather than a single blank line as before. This
should cause fewer issues when parsing the file.